// src/app/models/Chart.ts
export interface Chart {
    bestValue: string;
    configurationId: number;
    limit1: number;
    limit2: number;
    limit3: number;
    limit4: number;
    metricLabel: string;
    value: number;
    uomSymbol: string;
    uomPosition: string;
    bottomLabel: string;
    innerLabel: string;
    outerlabel: string;
  }